Simple Client-Server Application (Compiled)
This MP is a compilation of the features from ME1-ME5 where I need to create a full-text transfer program.

The following are the commands accepted by the program:
PWD – lists the current directory on the server
CD "directory name" – go down a specified directory (one level)
CD .. – go up one directory
LS – lists down the contents of the current directory (indicate directories)
GET <filename> - this reads a text file from the server and SAVES it on the client. Send an error message if the file is non-existent.