Simple Web Server (Full-Fledged)
This machine problem is the culmination of all MEs. This MP is now a Full-Fledged WebServer.

There is a simple webpage with 5 pictures which be requested by the browser. You should type the following url: http://127.0.0.1/index.html. MyServer responds to the request by sending the contents of the HTML file to the browser.